[batch] Add instance collections
