+++ batch/batch/driver/instance.py
@@ -13,16 +13,16 @@ log = logging.getLogger('instance')
 
 class Instance:
     @staticmethod
-    def from_record(app, record):
+    def from_record(app, inst_coll, record):
         return Instance(
-            app, record['name'], record['state'],
+            app, inst_coll, record['name'], record['state'],
             record['cores_mcpu'], record['free_cores_mcpu'],
             record['time_created'], record['failed_request_count'],
             record['last_updated'], record['ip_address'], record['version'],
             record['zone'])
 
     @staticmethod
-    async def create(app, name, activation_token, worker_cores_mcpu, zone):
+    async def create(app, inst_coll, name, activation_token, worker_cores_mcpu, zone):
         db = app['db']
 
         state = 'pending'
@@ -30,21 +30,20 @@ class Instance:
         token = secrets.token_urlsafe(32)
         await db.just_execute(
             '''
-INSERT INTO instances (name, state, activation_token, token, cores_mcpu, free_cores_mcpu, time_created, last_updated, version, zone)
-VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);
+INSERT INTO instances (name, state, activation_token, token, cores_mcpu, free_cores_mcpu, time_created, last_updated, version, zone, inst_coll)
+VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s);
 ''',
             (name, state, activation_token, token, worker_cores_mcpu,
-             worker_cores_mcpu, now, now, INSTANCE_VERSION, zone))
+             worker_cores_mcpu, now, now, INSTANCE_VERSION, zone, inst_coll.name))
         return Instance(
-            app, name, state, worker_cores_mcpu, worker_cores_mcpu, now,
+            app, inst_coll, name, state, worker_cores_mcpu, worker_cores_mcpu, now,
             0, now, None, INSTANCE_VERSION, zone)
 
-    def __init__(self, app, name, state, cores_mcpu, free_cores_mcpu,
+    def __init__(self, app, inst_coll, name, state, cores_mcpu, free_cores_mcpu,
                  time_created, failed_request_count, last_updated, ip_address,
                  version, zone):
         self.db = app['db']
-        self.instance_pool = app['inst_pool']
-        self.scheduler_state_changed = app['scheduler_state_changed']
+        self.inst_coll = inst_coll
         # pending, active, inactive, deleted
         self._state = state
         self.name = name
@@ -69,12 +68,12 @@ VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);
             'CALL activate_instance(%s, %s, %s);',
             (self.name, ip_address, timestamp))
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._state = 'active'
         self.ip_address = ip_address
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
-        self.scheduler_state_changed.set()
+        self.inst_coll.scheduler_state_changed.set()
 
         return rv['token']
 
@@ -93,13 +92,13 @@ VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);
             log.info(f'{self} with in-memory state {self._state} was already deactivated; {rv}')
             assert rv['cur_state'] in ('inactive', 'deleted')
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._state = 'inactive'
         self._free_cores_mcpu = self.cores_mcpu
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
         # there might be jobs to reschedule
-        self.scheduler_state_changed.set()
+        self.inst_coll.scheduler_state_changed.set()
 
     async def mark_deleted(self, reason, timestamp):
         if self._state == 'deleted':
@@ -115,18 +114,18 @@ VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s);
             log.info(f'{self} with in-memory state {self._state} could not be marked deleted; {rv}')
             assert rv['cur_state'] == 'deleted'
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._state = 'deleted'
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
     @property
     def free_cores_mcpu(self):
         return self._free_cores_mcpu
 
     def adjust_free_cores_in_memory(self, delta_mcpu):
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._free_cores_mcpu += delta_mcpu
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
     @property
     def failed_request_count(self):
@@ -166,10 +165,10 @@ WHERE name = %s;
 ''',
             (now, self.name))
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._failed_request_count = 0
         self._last_updated = now
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
     async def incr_failed_request_count(self):
         await self.db.execute_update(
@@ -179,9 +178,9 @@ SET failed_request_count = failed_request_count + 1 WHERE name = %s;
 ''',
             (self.name,))
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._failed_request_count += 1
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
     @property
     def last_updated(self):
@@ -193,9 +192,9 @@ SET failed_request_count = failed_request_count + 1 WHERE name = %s;
             'UPDATE instances SET last_updated = %s WHERE name = %s;',
             (now, self.name))
 
-        self.instance_pool.adjust_for_remove_instance(self)
+        self.inst_coll.adjust_for_remove_instance(self)
         self._last_updated = now
-        self.instance_pool.adjust_for_add_instance(self)
+        self.inst_coll.adjust_for_add_instance(self)
 
     def time_created_str(self):
         return time_msecs_str(self.time_created)
