+++ batch/batch/driver/templates/pool.html
@@ -1,0 +1,122 @@
+{% extends "layout.html" %}
+{% block title %}Pool{% endblock %}
+{% block content %}
+
+<h1>{{ pool.name }}</h1>
+
+<h2>Configuration</h2>
+<div class="attributes">
+  <form action="{{ base_path }}/config-update/pool/{{ pool.name }}" method="POST">
+    <div>Worker type: {{ pool.worker_type }}</div>
+    <div>Worker cores: <input name="worker_cores" value="{{ pool.worker_cores }}" /></div>
+    <div>Worker boot disk size (in GB): <input name="worker_disk_size_gb" value="{{ pool.worker_disk_size_gb }}" /></div>
+    <div>
+      <label for="worker_local_ssd_data_disk">Worker should use a Local SSD: </label>
+      <input type="checkbox"
+             id="worker_local_ssd_data_disk"
+             name="worker_local_ssd_data_disk"
+             {% if config['worker_local_ssd_data_disk']%}checked{% endif %}
+      value="true" />
+    </div>
+    <div>Worker PD SSD data disk size (in GB): <input name="worker_pd_ssd_data_disk_size_gb" value="{{ pool.worker_pd_ssd_data_disk_size_gb }}" /></div>
+    <div>
+      <label for="enable_standing_worker">Enable standing worker: </label>
+      <input type="checkbox"
+             id="enable_standing_worker"
+             name="enable_standing_worker"
+             {% if config['enable_standing_worker']%}checked{% endif %}
+      value="true" />
+    </div>
+    <div>Standing worker cores: <input name="standing_worker_cores" value="{{ pool.standing_worker_cores }}" /></div>
+    <div>Max instances: <input name="max_instances" value="{{ pool.max_instances }}" /></div>
+    <div>Max live instances: <input name="max_live_instances" value="{{ pool.max_live_instances }}" /></div>
+    <input type="hidden" name="_csrf" value="{{ csrf_token }}"/>
+    <button>
+      Update
+    </button>
+  </form>
+</div>
+
+<h2>User Resources</h2>
+<table class="data-table" id="user_resources">
+  <thead>
+  <tr>
+    <th>User</th>
+    <th>Ready Jobs</th>
+    <th>Ready Cores</th>
+    <th>Allocated Cores</th>
+    <th>Running Jobs</th>
+    <th>Running Cores</th>
+  </tr>
+  </thead>
+  <tbody>
+  {% for user in user_resources %}
+  <tr>
+    <td>{{ user['user'] }}</td>
+    <td class="numeric-cell">{{ user['n_ready_jobs'] }}</td>
+    <td class="numeric-cell">{{ user['ready_cores_mcpu'] / 1000 }}</td>
+    <td class="numeric-cell">{{ user['allocated_cores_mcpu'] / 1000 }}</td>
+    <td class="numeric-cell">{{ user['n_running_jobs'] }}</td>
+    <td class="numeric-cell">{{ user['running_cores_mcpu'] / 1000 }}</td>
+  </tr>
+  {% endfor %}
+  </tbody>
+</table>
+
+<h2>Status</h2>
+<div class="attributes">
+  <div>Ready cores: {{ ready_cores_mcpu / 1000 }}</div>
+</div>
+<table class="data-table" id="status">
+  <thead>
+  <tr>
+    <th>Pending</th>
+    <th>Active</th>
+    <th>Inactive</th>
+    <th>Deleted</th>
+    <th>Live Total Cores</th>
+    <th>Live Free Cores</th>
+  </tr>
+  </thead>
+  <tbody>
+  <tr>
+    <td class="numeric-cell">{{ pool.n_instances_by_state['pending'] }}</td>
+    <td class="numeric-cell">{{ pool.n_instances_by_state['active'] }}</td>
+    <td class="numeric-cell">{{ pool.n_instances_by_state['inactive'] }}</td>
+    <td class="numeric-cell">{{ pool.n_instances_by_state['deleted'] }}</td>
+    <td class="numeric-cell">{{ pool.live_total_cores_mcpu / 1000 }}</td>
+    <td class="numeric-cell">{{ pool.live_free_cores_mcpu / 1000 }}</td>
+  </tr>
+  </tbody>
+</table>
+
+<h2>Instances</h2>
+<table class="data-table" id="instances">
+  <thead>
+  <tr>
+    <th>Name</th>
+    <th>Zone</th>
+    <th>Version</th>
+    <th>State</th>
+    <th>Free Cores</th>
+    <th>Failed Requests</th>
+    <th>Time Created</th>
+    <th>Last Updated</th>
+  </tr>
+  </thead>
+  <tbody>
+  {% for instance in instances %}
+  <tr>
+    <td>{{ instance.name }}</td>
+    <td>{{ instance.zone }}</td>
+    <td class="numeric-cell">{{ instance.version }}</td>
+    <td>{{ instance.state }}</td>
+    <td class="numeric-cell">{{ instance.free_cores_mcpu / 1000 }} / {{ instance.cores_mcpu / 1000 }}</td>
+    <td class="numeric-cell">{{ instance.failed_request_count }}</td>
+    <td>{{ instance.time_created_str() }}</td>
+    <td>{{ instance.last_updated_str() }} ago</td>
+  </tr>
+  {% endfor %}
+  </tbody>
+</table>
+{% endblock %}
