+++ batch/batch/driver/templates/user_resources.html
@@ -9,7 +9,6 @@
         <th>Name</th>
         <th>Ready Jobs</th>
         <th>Ready Cores</th>
-        <th>Allocated Cores</th>
         <th>Running Jobs</th>
         <th>Running Cores</th>
     </tr>
@@ -20,7 +19,6 @@
         <td>{{ user['user'] }}</td>
         <td class="numeric-cell">{{ user['n_ready_jobs'] }}</td>
         <td class="numeric-cell">{{ user['ready_cores_mcpu'] / 1000 }}</td>
-        <td class="numeric-cell">{{ user['allocated_cores_mcpu'] / 1000 }}</td>
         <td class="numeric-cell">{{ user['n_running_jobs'] }}</td>
         <td class="numeric-cell">{{ user['running_cores_mcpu'] / 1000 }}</td>
     </tr>
