+++ batch/batch/driver/zone_monitor.py
@@ -1,10 +1,11 @@
 import asyncio
 import logging
+import random
 from hailtop import aiotools
 from hailtop.utils import retry_long_running, url_basename
 
 from ..utils import WindowFractionCounter
-from ..batch_configuration import BATCH_GCP_REGIONS
+from ..batch_configuration import BATCH_GCP_REGIONS, GCP_ZONE
 
 log = logging.getLogger('zone_monitor')
 
@@ -64,6 +65,28 @@ class ZoneMonitor:
 
     def shutdown(self):
         self.task_manager.shutdown()
+
+    def get_zone(self, worker_cores, worker_local_ssd_data_disk, worker_pd_ssd_data_disk_size_gb):
+        if self.app['inst_coll_manager'].global_live_total_cores_mcpu // 1000 < 1_000:
+            zone = GCP_ZONE
+        else:
+            zone_weights = self.zone_weights(worker_cores, worker_local_ssd_data_disk,
+                                             worker_pd_ssd_data_disk_size_gb)
+
+            if not zone_weights:
+                return None
+
+            zones = [zw.zone for zw in zone_weights]
+
+            zone_prob_weights = [
+                min(zw.weight, 10) * self.zone_success_rate.zone_success_rate(zw.zone)
+                for zw in zone_weights]
+
+            log.info(f'zone_success_rate {self.zone_success_rate}')
+            log.info(f'zone_prob_weights {zone_prob_weights}')
+
+            zone = random.choices(zones, zone_prob_weights)[0]
+        return zone
 
     def zone_weights(self, worker_cores, worker_local_ssd_data_disk, worker_pd_ssd_data_disk_size_gb):
         if not self.region_info:
