+++ batch/batch/front_end/validate.py
@@ -1,5 +1,6 @@
 from hailtop.batch_client.parse import (MEMORY_REGEX, MEMORY_REGEXPAT,
                                         CPU_REGEX, CPU_REGEXPAT)
+from ..globals import valid_worker_types
 
 from hailtop.utils.validate import bool_type, dictof, keyed, listof, int_type, nullable, \
     numeric, oneof, regex, required, str_type, switch, ValidationError
@@ -16,6 +17,7 @@ image_str = str_type
 # image -> process/image
 # mount_docker_socket -> process/mount_docker_socket
 # pvc_size -> resources/storage
+
 
 job_validator = keyed({
     'always_run': bool_type,
@@ -53,7 +55,8 @@ job_validator = keyed({
     'resources': keyed({
         'memory': regex(MEMORY_REGEXPAT, MEMORY_REGEX),
         'cpu': regex(CPU_REGEXPAT, CPU_REGEX),
-        'storage': regex(MEMORY_REGEXPAT, MEMORY_REGEX)
+        'storage': regex(MEMORY_REGEXPAT, MEMORY_REGEX),
+        'worker_type': oneof('standard', 'highcpu', 'highmem')
     }),
     'secrets': listof(keyed({
         required('namespace'): k8s_str,
