+++ batch/deployment.yaml
@@ -217,6 +217,8 @@ spec:
            value: "375M"
          - name: HAIL_BATCH_JOB_DEFAULT_STORAGE
            value: "1G"
+         - name: HAIL_BATCH_JOB_WORKER_TYPE
+           value: "standard"
 {% endif %}
 {% if deploy %}
          - name: HAIL_BATCH_BUCKET_NAME
